% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{constants}
\alias{constants}
\alias{FINITE_INF}
\alias{NSIM}
\title{Constants used in \pkg{superspreading}}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
FINITE_INF

NSIM
}
\description{
\code{FINITE_INF} is a large finite number used to approximate \code{Inf}.

\code{NSIM} is the number of simulations run when generating random samples or
branching process simulation replicates.
}
\keyword{datasets}
