% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offspring_distributions.R
\name{dpoisweibull}
\alias{dpoisweibull}
\title{Density of the poisson-Weibull compound distribution}
\usage{
dpoisweibull(x, shape, scale)
}
\arguments{
\item{x}{A \code{number} for the quantile of the distribution.}

\item{shape}{A \code{number} for the shape parameter of the distribution.}

\item{scale}{A \code{number} for the scale parameter of the distribution.}
}
\value{
A \code{numeric} vector of the density of the poisson-Weibull
distribution.
}
\description{
Density of the poisson-Weibull compound distribution
}
\details{
The function is vectorised so a vector of quantiles can be input
and the output will have an equal length.
}
\examples{
dpoisweibull(x = 10, shape = 1, scale = 2)
dpoisweibull(x = 1:10, shape = 1, scale = 2)
}
