% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cluster_size.R
\name{estimate_cluster_size}
\alias{estimate_cluster_size}
\title{Calculate the parameters of an offspring distribution given a cluster size(s)}
\usage{
estimate_cluster_size(x, R, k, offspring = "nbinom")
}
\arguments{
\item{x}{A vector of cluster sizes}

\item{R}{A \code{number} specifying the R parameter (i.e. average secondary cases
per infectious individual)}

\item{k}{A \code{number} specifying the  k parameter (i.e. overdispersion in
offspring distribution from fitted negative binomial)}

\item{offspring}{A specified offspring distribution}
}
\value{
A vector of named parameter values and the associated
maximum likleihood value
}
\description{
Calculate the parameters of an offspring distribution given a cluster size(s)
}
\examples{
cluster_sizes <- c(rep(1, 11), rep(2, 2), rep(3, 3), 4, rep(5, 2), 24)
estimate_cluster_size(x = cluster_sizes, R = 2, k = 0.5)
}
