% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_to_clusters.R
\name{fit_to_clusters}
\alias{fit_to_clusters}
\title{Example of a function to read stuff}
\usage{
fit_to_clusters(x)
}
\arguments{
\item{x}{the path to the file to read stuff from}
}
\description{
This function reads stuff into R. More details...
}
\examples{
some_path <- "path to your file"
read_stuff(some_path)
}
\author{
Adam Kucharski
}
