% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gamma}
\alias{gamma}
\alias{dgammaRk}
\alias{pgammaRk}
\alias{fvx}
\title{Defines the gamma functions in terms of the mean reproductive number
(R) and the dispersion parameter (k)}
\usage{
dgammaRk(x, R, k)

pgammaRk(x, R, k)

fvx(x, R, k)
}
\arguments{
\item{R}{A \code{number} specifying the \eqn{R} parameter (i.e. average
secondary cases per infectious individual).}

\item{k}{A \code{number} specifying the  \eqn{k} parameter (i.e. dispersion in
offspring distribution from fitted negative binomial).}
}
\description{
\itemize{
\item \code{\link[=dgammaRk]{dgammaRk()}} for the gamma density function
\item \code{\link[=pgammaRk]{pgammaRk()}} for the gamma distribution function
\item \code{\link[=fvx]{fvx()}} fore the gamma probability density function (pdf) describing the
individual reproductive number \eqn{\nu} given R0 and k
}
}
\keyword{internal}
