% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lseq}
\alias{lseq}
\title{Generates a log scaled sequence of real numbers}
\usage{
lseq(from, to, length.out)
}
\arguments{
\item{from, to}{the starting and (maximal) end values of the
    sequence.  Of length \code{1} unless just \code{from} is supplied as
    an unnamed argument.}

\item{length.out}{desired length of the sequence.  A
    non-negative number, which for \code{seq} and \code{seq.int} will be
    rounded up if fractional.}
}
\value{
\code{seq.int} and the default method of \code{seq} for numeric
  arguments return a vector of type \code{"integer"} or \code{"double"}:
  programmers should not rely on which.

  \code{seq_along} and \code{seq_len} return an integer vector, unless
  it is a \emph{\link[base]{long vector}} when it will be double.
}
\description{
Generates a log scaled sequence of real numbers
}
\keyword{internal}
