% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offspring_distributions.R
\name{ppoisweibull}
\alias{ppoisweibull}
\title{Cumulative distribution function of the poisson-Weibull compound
distribution}
\usage{
ppoisweibull(q, shape, scale)
}
\arguments{
\item{q}{A \code{number} for the quantile of the distribution}

\item{shape}{A \code{number} for the shape parameter of the distribution}

\item{scale}{A \code{number} for the scale parameter of the distribution}
}
\value{
A \code{numeric} vector of the distribution function.
}
\description{
Cumulative distribution function of the poisson-Weibull compound
distribution
}
\details{
The function is vectorised so a vector of quantiles can be input
and the output will have an equal length.
}
\examples{
ppoisweibull(q = 10, shape = 1, scale = 2)
ppoisweibull(q = 1:10, shape = 1, scale = 2)
}
