% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.ic_tbl}
\alias{print.ic_tbl}
\title{Print method for \verb{<ic_tbl>}}
\usage{
\method{print}{ic_tbl}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{x}, called for printing side-effects.
}
\description{
Print method for \verb{<ic_tbl>}
}
