% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_contain.R
\name{probability_contain}
\alias{probability_contain}
\title{Probability that an outbreak will be contained}
\usage{
probability_contain(
  R,
  k,
  num_init_infect = 1,
  control,
  control_type = c("population", "individual"),
  stochastic = TRUE,
  ...,
  case_threshold = 100,
  epidist
)
}
\arguments{
\item{R}{A \code{number} specifying the R parameter (i.e. average secondary cases
per infectious individual).}

\item{k}{A \code{number} specifying the  k parameter (i.e. overdispersion in
offspring distribution from fitted negative binomial).}

\item{num_init_infect}{A \code{count} specifying the number of initial infections.}

\item{control}{Control strength, 0 is no control measures, 1 is complete
control.}

\item{control_type}{Either \code{"population"} or \code{"individual"} for
population-level or individual-level control measures.}

\item{stochastic}{Whether to use a stochastic branching process model or the
probability of extinction.}

\item{...}{arguments to be passed to \code{\link[bpmodels:chain_sim]{bpmodels::chain_sim()}}.}

\item{case_threshold}{A number for the threshold of the number of cases below
which the epidemic is considered contained.}

\item{epidist}{An \verb{<epidist>} object. An S3 class for working with
epidemiological parameters/distributions, see \code{\link[epiparameter:epidist]{epiparameter::epidist()}}.}
}
\value{
A \code{number} for the probability of containment.
}
\description{
Containment is defined as the size of the transmission chain
not reaching the \code{case_threshold} (default = 100).
}
\examples{
probability_contain(R = 1.5, k = 0.5, control = 1)
}
\references{
Lloyd-Smith, J. O., Schreiber, S. J., Kopp, P. E., & Getz, W. M. (2005)
Superspreading and the effect of individual variation on disease emergence.
Nature, 438(7066), 355-359. \doi{10.1038/nature04153}
}
