% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_epidemic.R
\name{probability_extinct}
\alias{probability_extinct}
\title{Calculates the probability a branching process will go extinct based on
R, k and initial cases}
\usage{
probability_extinct(R, k, a)
}
\arguments{
\item{R}{A \code{number} specifying the R parameter (i.e. average secondary cases
per infectious individual)}

\item{k}{A \code{number} specifying the  k parameter (i.e. overdispersion in
offspring distribution from fitted negative binomial)}

\item{a}{A \code{count} specifying the number of initial infections}
}
\value{
A value with the probability of going extinct
}
\description{
Calculates the probability a branching process will not causes
an epidemic and will go extinct. This is the complement of the probability
of a disease causing an epidemic (\code{\link[=probability_epidemic]{probability_epidemic()}}).
}
\examples{
probability_extinct(R = 1.5, k = 0.1, a = 10)
}
\references{
Lloyd-Smith, J. O., Schreiber, S. J., Kopp, P. E., & Getz, W. M. (2005).
Superspreading and the effect of individual variation on disease emergence.
Nature, 438(7066), 355-359. \doi{10.1038/nature04153}
}
\seealso{
\code{\link[=probability_epidemic]{probability_epidemic()}}
}
