% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superspreading-package.R
\docType{package}
\name{superspreading-package}
\alias{superspreading}
\alias{superspreading-package}
\title{superspreading: Understand Individual-Level Variation in Infectious Disease Transmission}
\description{
Estimate and understand individual-level variation in transmission. Implements density and cumulative compound Poisson discrete distribution functions, as well as functions to calculate infectious disease outbreak statistics given epidemiological parameters on individual-level transmission; including the probability of an outbreak becoming an epidemic/extinct, or the cluster size statistics, e.g. what proportion of cases cause X\% of transmission.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/superspreading}
  \item \url{https://epiverse-trace.github.io/superspreading/}
  \item Report bugs at \url{https://github.com/epiverse-trace/superspreading/issues}
}

}
\author{
\strong{Maintainer}: Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Adam Kucharski \email{adam.kucharski@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8814-9421}{ORCID}) [copyright holder]
  \item Dillon C. Adam \email{dcadam@hku.hk} (\href{https://orcid.org/0000-0002-7485-9905}{ORCID})
}

Other contributors:
\itemize{
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-2842-3406}{ORCID}) (.chain_sim uses code from bpmodels::chain_sim) [contributor, copyright holder]
  \item Pratik Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID}) [reviewer]
  \item Hugo Gruson \email{hugo@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [reviewer]
  \item James M. Azam \email{james.azam@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5782-7330}{ORCID}) [reviewer, contributor]
  \item Chris Hartgerink \email{chris@data.org} (\href{https://orcid.org/0000-0003-1050-6809}{ORCID}) [reviewer]
}

}
\keyword{internal}
