% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_epidemic.R
\name{probability_epidemic}
\alias{probability_epidemic}
\title{Calculates the probability a branching process will cause an epidemic
(i.e. probability will fail to go extinct) based on R, k and initial cases}
\usage{
probability_epidemic(R, k, a)
}
\arguments{
\item{R}{A \code{number} specifying the R parameter (i.e. average secondary cases
per infectious individual)}

\item{k}{A \code{number} specifying the  k parameter (i.e. overdispersion in
offspring distribution from fitted negative binomial)}

\item{a}{A \code{count} specifying the number of initial infections}
}
\value{
A value with the probablity of a large epidemic
}
\description{
Calculates the probability a branching process will cause an epidemic
(i.e. probability will fail to go extinct) based on R, k and initial cases
}
\examples{
probability_epidemic(1.5, 0.1, 10)
}
