% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superspreading-package.R
\docType{package}
\name{superspreading-package}
\alias{superspreading}
\alias{superspreading-package}
\title{superspreading: Estimate Individual-Level Variation in Transmission}
\description{
Implements functions to estimate individual-level variation in transmission.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/superspreading}
  \item \url{https://epiverse-trace.github.io/superspreading/}
  \item Report bugs at \url{https://github.com/epiverse-trace/superspreading/issues}
}

}
\author{
\strong{Maintainer}: Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Adam Kucharski \email{adam.kucharski@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8814-9421}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-2842-3406}{ORCID}) [contributor]
  \item Pratik Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID}) [reviewer]
  \item Hugo Gruson \email{hugo.gruson@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [reviewer]
}

}
\keyword{internal}
